Move rescale_coordinates to _transform.py
It will be joined by the new rotate_coordinates function.
