+++ doc/api/index.rst
@@ -11,43 +11,51 @@ Coordinate generation
 ---------------------
 
 .. autosummary::
-   :toctree: generated/
+    :toctree: generated/
 
-   line_coordinates
-   random_coordinates
-   random_coordinates_spherical
-   grid_coordinates
-   profile_coordinates
-   great_circle_coordinates
+    line_coordinates
+    random_coordinates
+    random_coordinates_spherical
+    grid_coordinates
+    profile_coordinates
+    great_circle_coordinates
+
+Coordinate transformation
+-------------------------
+
+.. autosummary::
+    :toctree: generated/
+
+    rescale_coordinates
+    rotate_coordinates
 
 Regions and bounding boxes
 --------------------------
 
 .. autosummary::
-   :toctree: generated/
+    :toctree: generated/
 
-   inside
-   get_region
-   pad_region
-   rescale_coordinates
+    inside
+    get_region
+    pad_region
 
 Splitting points into blocks and windows
 ----------------------------------------
 
 .. autosummary::
-   :toctree: generated/
+    :toctree: generated/
 
-   block_split
-   expanding_window
-   rolling_window
-   rolling_window_spherical
+    block_split
+    expanding_window
+    rolling_window
+    rolling_window_spherical
 
 Other utilities
 ---------------
 
 .. autosummary::
-   :toctree: generated/
+    :toctree: generated/
 
-   neighbor_distance_statistics
-   spacing_to_size
-   shape_to_spacing
+    neighbor_distance_statistics
+    spacing_to_size
+    shape_to_spacing
